requirements.txt
mcp[cli]
pypdf
llama-parse
python-dotenv
nest-asyncio
neo4j>=5.0.0
Pillow
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart
cloud-sql-python-connector[pg8000]
google-cloud-storage
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
strands-agents
strands-amazon-nova
openai